+++ src/View/Components/Select/Searchable.php
@@ -16,8 +16,8 @@ class Searchable extends Styled
         public ?bool $multiple = false,
         public ?string $select = null,
         public ?array $selectable = [],
-        public ?array $after = null,
-        public ?array $before = null,
+        public ?string $after = null,
+        public ?string $before = null,
         private readonly bool $ignoreValidations = false,
     ) {
         parent::__construct(
